int vs. long problem resolved, all tests passing, submission tests passed

+++ src/six/ChallengeSix.java
@@ -1,7 +1,7 @@
 package six;
 
 public class ChallengeSix {
-    public static long calcSumOfSquares(int lastNum) {
+    public static long calcSumOfSquares(long lastNum) {
         return ((lastNum * (lastNum+1) * ((2*lastNum) + 1))/6);
     }
 
@@ -9,7 +9,7 @@ public class ChallengeSix {
         return (long) (((i / 2) * (1 + i)) * ((i / 2) * (1 + i)));
     }
 
-    public static long calcSumSquareDiff(int lastNum) {
+    public static long calcSumSquareDiff(long lastNum) {
         return Math.abs(calcSumOfSquares(lastNum) - calcSquareOfSums(lastNum));
     }
 }
